refactor(datagrid): cell-owned focus border replaces FocusOverlayView synchronization
