+++ TablePro/Views/Results/Extensions/DataGridView+Selection.swift
@@ -52,10 +52,6 @@ extension TableViewCoordinator {
         if keyTableView.focusedColumn != newFocus.column {
             keyTableView.focusedColumn = newFocus.column
         }
-        keyTableView.updateFocusOverlay()
-        DispatchQueue.main.async { [weak keyTableView] in
-            keyTableView?.updateFocusOverlay()
-        }
     }
 
     private func resolvedFocus(
